requirements.txt
# Python dependencies for the project

pandas>=1.3.0          # For data manipulation and reading CSV files
SQLAlchemy>=1.4.0      # For database interaction (ORM and Core)
mysql-connector-python>=8.0.0 # MySQL driver for SQLAlchemy
python-dotenv>=0.19.0  # For loading environment variables from .env file